% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdtotals}
\alias{birdtotals}
\title{Bird population totals in Sweden}
\format{A data frame [477 x 7]
\describe{
  \item{Art}{Art numeric code identifier}
  \item{Arthela}{The whole Swedish species name}
  \item{Ind}{Indicator}
  \item{YPctChg}{Yearly change in percent}
  \item{Significance}{Statistical significance where * is lower and *** is higher}
  \item{Series}{Which route was used for the inventory - Vinter, Sommar, Standard or Natt}
  \item{Suffix}{The suffix contains some special abbreviations used in Arthela}
  \item{Vernacular}{Vernacular or Popular name - the common name for the species, cleaned from any qualifier or suffix}
}}
\source{
\url{http://www.fageltaxering.lu.se/sites/default/files/files/Data/populationsindex.xls}
}
\usage{
birdtotals
}
\description{
A dataset containing bird population trend indices from Sweden
This dataset is from 2016-02-16
}
\keyword{datasets}

