% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdtrends}
\alias{birdtrends}
\title{Bird population trends in Sweden}
\format{A data frame [12,598 x 6]
\describe{
  \item{Art}{Species Number}
  \item{Arthela}{Species Name}
  \item{Year}{Year}
  \item{Measure}{Measure for index}
  \item{ts}{Timestamp for the year}
  \item{Series}{Series - route used - Winter, Summer, Standard or Night}
  \item{Suffix}{Suffix - qualifier - such as (S) for South, (N) for North, ob for not determined etc}
  \item{Vernacular}{Vernacular or Popular name - the common name for the species, cleaned from any qualifier or suffix}
  \item{Scientific}{Scientific name, the latin binomen}
}}
\source{
\url{http://www.fageltaxering.lu.se/sites/default/files/files/Data/populationsindex.xls}
}
\usage{
birdtrends
}
\description{
A dataset containing bird population trend indices from Sweden
This dataset is from 2016-02-16
}
\keyword{datasets}

