% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birduuids}
\alias{birduuids}
\title{Bird identifiers mapped to unique identifiers used at The Naturalist / Naturforskaren (naturforskaren.se)}
\format{A data frame [22,840 x 2]
\describe{
  \item{species_uuid}{unique identifier for a species used in The Naturalist}
  \item{vernacular}{vernacular name in lower case}
  \item{sciname}{Scientific name}
}}
\source{
\url{http://naturforskaren.se}
}
\usage{
birduuids
}
\description{
A dataset containing bird population trend indices from Sweden
This dataset is from 2016-02-16
}
\keyword{datasets}

