% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R
\name{dy}
\alias{dy}
\title{Create a dynamic line graph (dygraph) for a specific bird species}
\usage{
dy(df, species, showgrid = FALSE)
}
\arguments{
\item{df}{a data frame structured like the bundled birdtrends dataset}

\item{species}{a character vector of species names}

\item{showgrid}{boolean flag to indicate whether the graph should use gridlines}
}
\description{
Create a dynamic line graph (dygraph) for a specific bird species
}

