% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naturalist.R
\name{get_nf_media_urls}
\alias{get_nf_media_urls}
\title{Retrieve image url from The Naturalist given latin binomen for a bird}
\usage{
get_nf_media_urls(sciname)
}
\arguments{
\item{sciname}{the scientific name for a bird species}
}
\value{
a character vector with image urls for the species
}
\description{
Retrieve image url from The Naturalist given latin binomen for a bird
}
\examples{
get_nf_media_urls("Erithacus rubecula")
}

