% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naturalist.R
\name{plot_image_url}
\alias{plot_image_url}
\title{Plot image from url}
\usage{
plot_image_url(url, width = 120L)
}
\arguments{
\item{url}{the image url}

\item{width}{the default width}
}
\value{
a plot with the image
}
\description{
Plot image from url
}
\examples{
plot_image_url(get_nf_media_urls("Turdus merula")[1])
}

