% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R
\name{plot_sbt_dy}
\alias{plot_sbt_dy}
\title{Plot a trend for a specific Swedish bird species using an dynamic graph}
\usage{
plot_sbt_dy(species, df = birdtrends)
}
\arguments{
\item{species}{a character vector with the vernacular name of the species}

\item{df}{optional data frame with bird trend data, defaults to the bundled birdtrends}
}
\value{
a dygraph - dynamic graph
}
\description{
Plot a trend for a specific Swedish bird species using an dynamic graph
}
\examples{
plot_sbt_dy("Häger")
}

