% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R
\name{plot_sbt_static}
\alias{plot_sbt_static}
\title{Plot a trend for a specific Swedish bird species using an static graph}
\usage{
plot_sbt_static(species, df = birdtrends, loess = FALSE,
  series = c("Vinter", "Sommar", "Standard", "Natt"), showlegend = TRUE,
  theme = theme_few())
}
\arguments{
\item{species}{a character vector with the vernacular name of the species}

\item{df}{optional data frame with bird trend data, defaults to the bundled birdtrends}

\item{loess}{boolean indicating whether to use loess smoothing for the trend}

\item{series}{a character vector with the routes to include, by 
default Winter, Summer, Standard and Night routes}

\item{showlegend}{boolean to show the color legend, by default TRUE}

\item{theme}{a ggplot theme from the ggtheme package, by default theme_few()}
}
\value{
a ggplot - static graph
}
\description{
Plot a trend for a specific Swedish bird species using an static graph
}
\examples{
plot_sbt_static("Häger")
plot_sbt_static("Havsörn", loess = TRUE)
plot_sbt_static("Duvhök", loess = TRUE, series = "Vinter")
}

