% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swedishbirdtrends-package.R
\docType{package}
\name{swedishbirdtrends-package}
\alias{swedishbirdtrends}
\alias{swedishbirdtrends-package}
\title{swedishbirdtrends - An R package distributing Swedish Bird Population Trends data and diagrams}
\description{
The swedishbirdtrends package provides tools for working with Swedish Bird Population Trends index data
and diagrams through R.
}
\details{
The data included is provided from The Swedish Bird Inventory at http://www.fageltaxering.lu.se at 2016-02-14.

The indices are calculated using TRIM - a method for providing indicators for time series of counts with many missing values.
}
\examples{
#see functions.
}
\references{
http://www.cso.cz/wpimages/other/vanstrien.pdf

http://www.cbs.nl/en-GB/menu/themas/natuur-milieu/methoden/trim/achtergrond-methode-trim.htm

http://www.cbs.nl/NR/rdonlyres/08046BFB-A8B4-49F8-A257-22BAE3606CC2/0/2000loglinearp033art.pdf
}
\keyword{package}

