% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendy.R
\name{turn}
\alias{turn}
\title{Turn "thin" bird index data into series spread across columns}
\usage{
turn(df, species)
}
\arguments{
\item{df}{a dataset formatted like the bundled birdtrends dataset}

\item{species}{a character vector of species names}
}
\value{
a data frame dplyr style
}
\description{
Turn "thin" bird index data into series spread across columns
}

